// src/app/model/cnh.model.ts
export class Cnh {
  dataPrimeiraHabilitacao: string;
  cpf: string;
  emissor: string;
  dataValidade: string;
  numeroDocIdentidadeEmissorUf: string;
  ufExpedicao: string;
  numeroRegistro: string;
  dataNascimento: string;
  nome: string;
  dataEmissao: string;
  nomePai: string;
  nomeMae: string;
  numeroRegistroRg: string;
  orgaoEmissorIdentidade: string;
  ufExpedicaoIdentidade: string;
  numeroFormulario: string;
  numeroRenach: string;
  numeroSeguranca: string;
  marcaDocumentoPrincipal: string;
  identificador: string;
}