// src/app/model/carteira-identidade.model.ts
export class CarteiraIdentidade {
    // identificador_cliente: string;
    nomeMae: string;
    nomePai: string;
    naturalidadeCidade: string;
    naturalidadeUf: string;
    nome: string;
    dataNascimento: string;
    dataEmissao: string;
    numeroRegistro: string;
    ufEmissor: string;
    emissor: string;
    marcaDocumentoPrincipal: string;
    cpf: string;
}